Using qint32 instead of qint64 for dir size to reduce the size of vector's elements twice. It must make the program more effective in the sense of space and cache errors

+++ DirSize.cpp
@@ -5,10 +5,10 @@
 #include <QDir>
 #include <QDebug>
 
-qint64 dirSize(const QString &dirPath)
+qint32 dirSize(const QString &dirPath)
 {
     QFileInfo info(dirPath);
-    qint64 size = 0;
+    qint32 size = 0;
     if(info.isDir())
     {
         QDir dir(dirPath);
